Struct paho_mqtt::client_persistence::ClientPersistenceBridge

pub struct ClientPersistenceBridge;

An empty struct used to collect the persistence callback funtions from the C library. These functions receive the persistence callbacks from the C library and then pass them on to the user-supplied struct which implements the ClientPersistence trait.

Methods

impl ClientPersistenceBridge

unsafe extern "C" fn on_open(
handle: *mut *mut c_void,
client_id: *const c_char,
server_uri: *const c_char,
context: *mut c_void
) -> c_int

unsafe extern "C" fn on_close(handle: *mut c_void) -> c_int

unsafe extern "C" fn on_put(
handle: *mut c_void,
key: *mut c_char,
bufcount: c_int,
buffers: *mut *mut c_char,
buflens: *mut c_int
) -> c_int

unsafe extern "C" fn on_get(
handle: *mut c_void,
key: *mut c_char,
buffer: *mut *mut c_char,
buflen: *mut c_int
) -> c_int

unsafe extern "C" fn on_remove(handle: *mut c_void, key: *mut c_char) -> c_int

unsafe extern "C" fn on_keys(
handle: *mut c_void,
keys: *mut *mut *mut c_char,
nkeys: *mut c_int
) -> c_int

unsafe extern "C" fn on_clear(handle: *mut c_void) -> c_int

unsafe extern "C" fn on_contains_key(
handle: *mut c_void,
key: *mut c_char
) -> c_int