Struct paho_mqtt::message::MessageBuilder

pub struct MessageBuilder { /* fields omitted */ }

Builder to create a new Message

Methods

impl MessageBuilder

fn new() -> MessageBuilder

Create a new message builder.

fn topic(self, topic: &str) -> MessageBuilder

Sets the topic for the message

Arguments

topic The topic on which the message should be published.

fn payload<V>(self, payload: V) -> MessageBuilder where
V: Into<Vec<u8>>,

Sets the payload for the message

Arguments

payload The binary payload of the message

fn qos(self, qos: i32) -> MessageBuilder

Sets the Quality of Service for the message.

Arguments

qos The quality of service for the message.

fn retained(self, retained: bool) -> MessageBuilder

Sets whether or not the published message should be retained by the broker.

Arguments

retained Set true if the message should be retained by the broker, false if not.

fn finalize(self) -> Message

Finalize the builder to create the message.